requirements.txt
Django==4.2.8
django-admin-sortable2==2.1.10
environs==10.0.0
pillow==10.2.0
